fix: get first still after fragment start + switch to post request
